move load_tree function
